% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approxdens.R
\name{approxdens}
\alias{approxdens}
\title{Empirical Density Function}
\usage{
approxdens(x, ...)
}
\arguments{
\item{x}{\link[base]{numeric} \link[base]{vector}, observations}

\item{...}{additional parameters of \link[stats]{density.default}}
}
\value{
The function [approxdens()] returns a \link[base]{function}.
}
\description{
..
}
\details{
\link[stats]{approx} inside \link[stats]{density.default}

another 'layer' of \link[stats]{approxfun}
}
\examples{
set.seed(135); x = rnorm(1e3L)
(f = approxdens(x))
x[1:3] |> f()
}
\keyword{internal}
