% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.fmx.R
\name{as.fmx.fitdist}
\alias{as.fmx.fitdist}
\title{Convert \link[fitdistrplus]{fitdist} Objects to \linkS4class{fmx} Class}
\usage{
\method{as.fmx}{fitdist}(x, ...)
}
\arguments{
\item{x}{\link[fitdistrplus]{fitdist} object}

\item{...}{..}
}
\value{
Function \code{\link[=as.fmx.fitdist]{as.fmx.fitdist()}} returns an \linkS4class{fmx} object.
}
\description{
To convert \link[fitdistrplus]{fitdist} objects (from package \CRANpkg{fitdistrplus})
to \linkS4class{fmx} class.
}
\keyword{internal}
