% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.fmx.R
\name{as.fmx}
\alias{as.fmx}
\title{Turn Various Objects to \linkS4class{fmx} Class}
\usage{
as.fmx(x, ...)
}
\arguments{
\item{x}{an R object}

\item{...}{additional parameters, see **Arguments** in individual S3 dispatches}
}
\value{
S3 generic function [as.fmx()] returns an \linkS4class{fmx} object.
}
\description{
Turn various objects created in other R packages 
to \linkS4class{fmx} class.
}
\details{
Various mixture distribution estimates obtained from other R packages
are converted to \linkS4class{fmx} class, 
so that we could take advantage of all methods defined for \linkS4class{fmx} objects.
}
