% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.fmx.R
\name{as.fmx.Skew.normal}
\alias{as.fmx.Skew.normal}
\title{Convert \code{Skew.normal} Object to \linkS4class{fmx}}
\usage{
\method{as.fmx}{Skew.normal}(x, data, ...)
}
\arguments{
\item{x}{\code{'Skew.normal'} object,
returned from \link[mixsmsn]{smsn.mix} with parameter
\code{family = 'Skew.normal'}.}

\item{data}{\link[base]{numeric} \link[base]{vector}}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \code{\link[=as.fmx.Skew.normal]{as.fmx.Skew.normal()}} returns an \linkS4class{fmx} object.
}
\description{
To convert \code{Skew.normal} object (from package \CRANpkg{mixsmsn})
to \linkS4class{fmx} class.
}
\note{
\link[mixsmsn]{smsn.mix} does not offer a parameter to keep the input data, as of 2021-10-06.
}
\keyword{internal}
