% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{coef.fmx}
\alias{coef.fmx}
\title{Parameter Estimates of \linkS4class{fmx} object}
\usage{
\method{coef}{fmx}(object, internal = FALSE, ...)
}
\arguments{
\item{object}{\linkS4class{fmx} object}

\item{internal}{\link[base]{logical} scalar, either for the user-friendly parameters (`FALSE`, default)
(e.g., `mean,sd` for normal mixture, and `A,B,g,h` for Tukey \eqn{g}-and-\eqn{h} mixture), or
for the internal/unconstrained parameters (`TRUE`).}

\item{...}{place holder for S3 naming convention}
}
\value{
The function [coef.fmx()] returns a \link[base]{numeric} \link[base]{vector}.
}
\description{
..
}
\details{
The function [coef.fmx()] returns the estimates of the user-friendly parameters (`parm = 'user'`), 
or the internal/unconstrained parameters (\code{parm = 'internal'}).
When the distribution has constraints on one or more parameters, 
function [coef.fmx()] does not return the estimates (which is constant \code{0}) of the constrained parameters.
}
\keyword{internal}
