% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{confint.fmx}
\alias{confint.fmx}
\title{Confidence Interval of \linkS4class{fmx} Object}
\usage{
\method{confint}{fmx}(object, ..., level = 0.95)
}
\arguments{
\item{object}{\linkS4class{fmx} object}

\item{...}{place holder for S3 naming convention}

\item{level}{confidence level, default \eqn{95\%}.}
}
\value{
The function [confint.fmx()] returns a \link[base]{matrix}
}
\description{
...
}
\details{
The function [confint.fmx()] returns the Wald-type confidence intervals based on the user-friendly parameters (`parm = 'user'`),
 or the internal/unconstrained parameters (`parm = 'internal'`).
When the distribution has constraints on one or more parameters, 
function [confint.fmx()] does not return the confident intervals of for the constrained parameters.
}
\keyword{internal}
