% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmx2dbl.R
\name{dbl2fmx}
\alias{dbl2fmx}
\title{Inverse of [fmx2dbl], for internal use}
\usage{
dbl2fmx(x, K, distname, ...)
}
\arguments{
\item{x}{\link[base]{numeric} \link[base]{vector}, unrestricted parameters}

\item{K}{\link[base]{integer} scalar}

\item{distname}{\link[base]{character} scalar}

\item{...}{additional parameters, not currently used}
}
\value{
The function [dbl2fmx()] returns a \link[base]{list} with two elements `$pars` and `$w`
}
\description{
..
}
\details{
Only used in downstream function `QuantileGH::QLMDe` and unexported function `QuantileGH:::qfmx_gr`, not compute intensive.
}
\keyword{internal}
