% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqr.R
\name{dfmx}
\alias{dfmx}
\alias{pfmx}
\alias{qfmx}
\alias{rfmx}
\title{Density, Distribution and Quantile of Finite Mixture Distribution}
\usage{
dfmx(
  x,
  dist,
  distname = dist@distname,
  K = dim(pars)[1L],
  pars = dist@pars,
  w = dist@w,
  ...,
  log = FALSE
)

pfmx(
  q,
  dist,
  distname = dist@distname,
  K = dim(pars)[1L],
  pars = dist@pars,
  w = dist@w,
  ...,
  lower.tail = TRUE,
  log.p = FALSE
)

qfmx(
  p,
  dist,
  distname = dist@distname,
  K = dim(pars)[1L],
  pars = dist@pars,
  w = dist@w,
  interval = qfmx_interval(dist = dist),
  ...,
  lower.tail = TRUE,
  log.p = FALSE
)

rfmx(
  n,
  dist,
  distname = dist@distname,
  K = dim(pars)[1L],
  pars = dist@pars,
  w = dist@w
)
}
\arguments{
\item{x, q}{\link[base]{numeric} \link[base]{vector}, quantiles, `NA_real_` value(s) allowed.}

\item{dist}{\linkS4class{fmx} object, a finite mixture distribution}

\item{distname, K, pars, w}{auxiliary parameters, whose default values are determined by argument `dist`.
The user-specified \link[base]{vector} of `w` does not need to sum up to 1; `w/sum(w)` will be used internally.}

\item{...}{additional parameters}

\item{log, log.p}{\link[base]{logical} scalar. 
If `TRUE`, probabilities are given as \eqn{\log(p)}.}

\item{lower.tail}{\link[base]{logical} scalar. 
If `TRUE` (default), probabilities are \eqn{Pr(X\le x)}, otherwise, \eqn{Pr(X>x)}.}

\item{p}{\link[base]{numeric} \link[base]{vector}, probabilities.}

\item{interval}{\link[base]{length}-2 \link[base]{numeric} \link[base]{vector}, interval for root finding, see \link[rstpm2]{vuniroot}}

\item{n}{\link[base]{integer} scalar, number of observations.}
}
\value{
The function [dfmx()] returns a \link[base]{numeric} \link[base]{vector} of probability density values of an \linkS4class{fmx} object at specified quantiles `x`.

The function [pfmx()] returns a \link[base]{numeric} \link[base]{vector} of cumulative probability values of an \linkS4class{fmx} object at specified quantiles `q`.

The function [qfmx()] returns an unnamed \link[base]{numeric} \link[base]{vector} of quantiles of an \linkS4class{fmx} object, based on specified cumulative probabilities `p`.

The function [rfmx()] generates random deviates of an \linkS4class{fmx} object.
}
\description{
Density function, distribution function, quantile function and random generation for a finite mixture distribution 
with normal or Tukey \eqn{g}-&-\eqn{h} components.
}
\details{
A computational challenge in function [dfmx()] is when mixture density is very close to 0,
which happens when the per-component log densities are negative with big absolute values.  
In such case, we cannot compute the log densities (i.e., `-Inf`).

The function [qfmx()] gives the quantile function, by numerically solving [pfmx].
One major challenge when dealing with the finite mixture of Tukey \eqn{g}-&-\eqn{h} family distribution
is that Brent–Dekker's method needs to be performed in both \link[TukeyGH77]{pGH} and [qfmx] functions, 
i.e. *two layers* of root-finding algorithm.
}
\note{
The function \link[stats]{qnorm} returns an unnamed \link[base]{vector} of quantiles, 
although \link[stats]{quantile} returns a named \link[base]{vector} of quantiles.
}
\keyword{internal}
