% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_ext.R
\name{dist_logtrans}
\alias{dist_logtrans}
\title{Distribution Parameters that needs to have a log-transformation}
\usage{
dist_logtrans(distname)
}
\arguments{
\item{distname}{\link[base]{character} scalar, name of distribution}
}
\value{
The function [dist_logtrans()] returns an \link[base]{integer} scalar
}
\description{
..
}
\keyword{internal}
