% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0S4.R
\docType{class}
\name{fmx-class}
\alias{fmx-class}
\title{\linkS4class{fmx} Class: Finite Mixture Parametrization}
\description{
An S4 object to specify 
the parameters and type of distribution 
of a one-dimensional finite mixture distribution.
}
\section{Slots}{

\describe{
\item{\code{distname}}{\link[base]{character} scalar, 
name of parametric distribution of the mixture components.
Currently, normal (`'norm'`) and Tukey \eqn{g}-&-\eqn{h} (`'GH'`) distributions are supported.}

\item{\code{pars}}{\link[base]{double} \link[base]{matrix}, 
all distribution parameters in the mixture. 
Each row corresponds to one component. Each column includes the same parameters of all components.
The order of rows corresponds to the (non-strictly) increasing order of the component location parameters.
The columns match the formal arguments of the corresponding distribution, 
e.g., `'mean'` and `'sd'` for \link[stats]{Normal} mixture, 
or `'A'`, `'B'`, `'g'` and `'h'` for Tukey \eqn{g}-&-\eqn{h} mixture.}

\item{\code{w}}{\link[base]{numeric} \link[base]{vector} of mixing proportions that must sum to 1}

\item{\code{data}}{(optional) \link[base]{numeric} \link[base]{vector}, the one-dimensional observations}

\item{\code{data.name}}{(optional) \link[base]{character} scalar, a human-friendly name of the observations}

\item{\code{vcov_internal}}{(optional) variance-covariance \link[base]{matrix} of the internal (i.e., unconstrained) estimates}

\item{\code{vcov}}{(optional) variance-covariance \link[base]{matrix} of the mixture distribution (i.e., constrained) estimates}

\item{\code{dist.ks}}{(optional) \link[base]{double} scalar, Kolmogorov-Smirnov distance, via \link[stats]{ks.test}}

\item{\code{dist.cvm}}{(optional) \link[base]{double} scalars, Cramer von Mises distance, via \link[goftest]{cvm.test}}

\item{\code{dist.kl}}{(optional) \link[base]{double} scalars, Kullback-Leibler distance}

\item{\code{logd}}{(optional) \link[base]{double} \link[base]{vector}, point-wise \link[base]{log}-density}

\item{\code{logLik}}{(optional) \link[stats]{logLik} object, log-likelihood}
}}

