% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmx_constraint.R
\name{fmx_constraint}
\alias{fmx_constraint}
\title{Parameter Constraint(s) of Mixture Distribution}
\usage{
fmx_constraint(
  dist,
  distname = dist@distname,
  K = dim(dist@pars)[1L],
  pars = dist@pars
)
}
\arguments{
\item{dist}{(optional) \linkS4class{fmx} object}

\item{distname}{\link[base]{character} scalar, name of distribution (see \linkS4class{fmx}),
default value determined by `dist`}

\item{K}{\link[base]{integer} scalar, number of components,
default value determined by `dist`}

\item{pars}{\link[base]{double} \link[base]{matrix}, 
distribution parameters of a finite mixture distribution (see \linkS4class{fmx}),
default value determined by `dist`}
}
\value{
The function [fmx_constraint()] returns the indices of internal parameters 
(only applicable to Tukey \eqn{g}-&-\eqn{h} mixture distribution, yet) to be constrained, 
based on the input \linkS4class{fmx} object `dist`.
}
\description{
Determine the parameter constraint(s) of a finite mixture distribution \linkS4class{fmx}, 
either by the value of parameters of such mixture distribution, 
or by a user-specified string.
}
\keyword{internal}
