% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmx_constraint.R
\name{getTeX}
\alias{getTeX}
\title{TeX Label (of Parameter Constraint(s)) of \linkS4class{fmx} Object}
\usage{
getTeX(dist, print_K = FALSE)
}
\arguments{
\item{dist}{\linkS4class{fmx} object}

\item{print_K}{\link[base]{logical} scalar, whether to print the number of components \eqn{K}.
Default `FALSE`.}
}
\value{
The function [getTeX()] returns a \link[base]{character} scalar 
(of TeX expression) of the constraint, 
primarily intended for end-users in plots.
}
\description{
Create TeX label of (parameter constraint(s)) of \linkS4class{fmx} object
}
\keyword{internal}
