% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kolmogorov_dist.R
\name{Kolmogorov_dist}
\alias{Kolmogorov_dist}
\title{One-Sample Kolmogorov Distance}
\usage{
Kolmogorov_dist(x, null, alternative = c("two.sided", "less", "greater"), ...)
}
\arguments{
\item{x}{\link[base]{numeric} \link[base]{vector}, observations \eqn{x}}

\item{null}{cumulative distribution \link[base]{function}}

\item{alternative}{\link[base]{character} scalar,
alternative hypothesis, either \code{'two.sided'} (default), \code{'less'}, or \code{'greater'}}

\item{...}{additional arguments of \code{null}}
}
\value{
Function \code{\link[=Kolmogorov_dist]{Kolmogorov_dist()}} returns a \link[base]{numeric} scalar.
}
\description{
To calculate the one-sample Kolmogorov distance between observations and
a distribution.
}
\details{
Function \code{\link[=Kolmogorov_dist]{Kolmogorov_dist()}} is different from \link[stats]{ks.test} in the
following aspects
\itemize{
\item {Ties in observations are supported.
The step function of empirical distribution is inspired by \link[stats]{ecdf}.
This is superior than \code{(0:(n - 1))/n} in \link[stats]{ks.test}.}
\item {Discrete distribution (with discrete observation) is supported.}
\item {Discrete distribution (with continuous observation) is not supported yet.
This will be an easy modification in future.}
\item {Only the one-sample Kolmogorov distance, not the one-sample Kolmogorov test, is returned,
to speed up the calculation.}
}
}
\examples{
set.seed(24); x1 = rnorm(n = 50L, mean = 2)
x1 |> stats::ks.test(y = pgamma, shape = 3, rate = 2)
x1 |> Kolmogorov_dist(null = pgamma, shape = 3, rate = 2) # exactly the same

set.seed(76); x2 = rnbinom(n = 1e2L, size = 500, prob = .4)
x2 |> stats::ks.test(y = pnbinom, size = 500, prob = .4) |> suppressWarnings() # warning on ties
x2 |> Kolmogorov_dist(null = pnbinom, size = 500, prob = .4) # not exactly the same
}
\keyword{internal}
