% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitdistrplus.R
\name{logLik.fitdist}
\alias{logLik.fitdist}
\title{Log-Likelihood of \link[fitdistrplus]{fitdist} Object}
\usage{
\method{logLik}{fitdist}(object, ...)
}
\arguments{
\item{object}{\link[fitdistrplus]{fitdist} object}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \code{\link[=logLik.fitdist]{logLik.fitdist()}} returns a \link[stats]{logLik} object, which
could be further used by \link[stats]{AIC} and \link[stats]{BIC}.

(I have written to the authors)
}
\description{
..
}
\details{
Output of \link[fitdistrplus]{fitdist} has elements \verb{$loglik}, \verb{$aic} and \verb{$bic},
but they are simply \link[base]{numeric} scalars.
\code{fitdistrplus:::logLik.fitdist} simply returns these elements.
}
\keyword{internal}
