% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtools.R
\name{mixEM_pars}
\alias{mixEM_pars}
\title{Names of Distribution Parameters of \code{'mixEM'} Object}
\usage{
mixEM_pars(object)
}
\arguments{
\item{object}{\code{'mixEM'} object, currently only the returned value of
\link[mixtools]{normalmixEM} and \link[mixtools]{gammamixEM} are supported}
}
\value{
Function \code{\link[=mixEM_pars]{mixEM_pars()}} returns a \link[base]{character} \link[base]{vector}
}
\description{
Names of distribution parameters of \code{'mixEM'} object, based on \CRANpkg{mixtools} 2020-02-05.
}
\seealso{
\link[mixtools]{normalmixEM} \link[mixtools]{gammamixEM}
}
\keyword{internal}
