% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlogis.R
\name{mlogis}
\alias{mlogis}
\alias{qmlogis}
\alias{pmlogis}
\title{Multinomial Probabilities & Logits}
\usage{
qmlogis(p)

pmlogis(q)
}
\arguments{
\item{p}{\link[base]{double} \link[base]{vector}, multinomial probabilities}

\item{q}{\link[base]{double} \link[base]{vector}, multinomial logits}
}
\value{
The function [pmlogis()] returns a \link[base]{double} \link[base]{vector} of multinomial probabilities \eqn{p}.

The function [qmlogis()] returns a \link[base]{double} \link[base]{vector} of multinomial logits \eqn{q}.
}
\description{
Transformation between the \link[base]{vector}s of multinomial probabilities and logits.
}
\details{
The function [pmlogis()] takes a length \eqn{k-1} \link[base]{double} \link[base]{vector} of 
multinomial logits \eqn{q} and convert them into length \eqn{k} multinomial probabilities \eqn{p}, 
regarding the *first* category as reference.

The function [qmlogis()] takes a length \eqn{k} \link[base]{double} \link[base]{vector} of 
multinomial probabilities \eqn{p} and convert them into length \eqn{k-1} multinomial logits \eqn{q}, 
regarding the *first* category as reference.
}
\note{
This transformation is a generalization of functions \link[stats]{plogis} (i.e., logit to probability)
and \link[stats]{qlogis} (i.e., probability to logit).
}
\examples{
c(2,5,3) |> qmlogis() |> pmlogis()

# various exceptions
c(1, 0) |> qmlogis() |> pmlogis()
c(0, 1) |> qmlogis() |> pmlogis()
c(0, 0, 1) |> qmlogis() |> pmlogis()
c(1, 0, 0, 0) |> qmlogis() |> pmlogis()
c(0, 1, 0, 0) |> qmlogis() |> pmlogis()
c(0, 0, 1, 0) |> qmlogis() |> pmlogis()
   
}
\keyword{internal}
