% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment.R
\name{moment_fmx}
\alias{moment_fmx}
\title{Moment of Each Component in an \linkS4class{fmx} Object}
\usage{
moment_fmx(object)
}
\arguments{
\item{object}{an \linkS4class{fmx} object}
}
\value{
The function [moment_fmx()] returns an object of the S4 \link[param2moment]{moment-class}.
}
\description{
To find moments of each component in an \linkS4class{fmx} object.
}
\details{
The function [moment_fmx()] calculates the moments 
and distribution characteristics of each mixture component of 
an S4 \linkS4class{fmx} object.
}
\keyword{internal}
