% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitdistrplus.R
\name{nobs.fitdist}
\alias{nobs.fitdist}
\title{Number of Observations in \link[fitdistrplus]{fitdist} Object}
\usage{
\method{nobs}{fitdist}(object, ...)
}
\arguments{
\item{object}{\link[fitdistrplus]{fitdist} object}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{nobs.fitdist} returns an \link[base]{integer} scalar
}
\description{
..
}
