% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmx_constraint.R
\name{npar.fmx}
\alias{npar.fmx}
\title{Number of Parameters of \linkS4class{fmx} Object}
\usage{
npar.fmx(dist)
}
\arguments{
\item{dist}{\linkS4class{fmx} object}
}
\value{
The function [npar.fmx()] returns an \link[base]{integer} scalar.
}
\description{
..
}
\details{
Also the degree-of-freedom in \link[stats]{logLik},
as well as `stats:::AIC.logLik` and `stats:::BIC.logLik`
}
\keyword{internal}
