% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpqr.R
\name{qfmx_interval}
\alias{qfmx_interval}
\title{Obtain `interval` for \link[rstpm2]{vuniroot} for Function [qfmx()]}
\usage{
qfmx_interval(
  dist,
  p = c(1e-06, 1 - 1e-06),
  distname = dist@distname,
  K = dim(pars)[1L],
  pars = dist@pars,
  w = dist@w,
  ...
)
}
\arguments{
\item{dist}{\linkS4class{fmx} object}

\item{p}{\link[base]{length}-2 \link[base]{numeric} \link[base]{vector}}

\item{distname, K, pars, w}{(optional) ignored if `dist` is provided}

\item{...}{additional parameters, currently not used}
}
\value{
The function [qfmx_interval()] returns a \link[base]{length}-2 \link[base]{numeric} \link[base]{vector}.
}
\description{
Obtain `interval` for \link[rstpm2]{vuniroot} for Function [qfmx()]
}
\keyword{internal}
