% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_fmx.R
\name{[,fmx,ANY,ANY,ANY-method}
\alias{[,fmx,ANY,ANY,ANY-method}
\title{Subset of Components in \linkS4class{fmx} Object}
\usage{
\S4method{[}{fmx,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{\linkS4class{fmx} object}

\item{i}{\link[base]{integer} or \link[base]{logical} \link[base]{vector},
the row indices of \emph{components} to be chosen, see \link[base]{[}}
}
\value{
An \linkS4class{fmx} object consisting of a subset of components.
information about the observations (e.g. slots \verb{@data} and \verb{@data.name}),
will be lost.
}
\description{
Taking subset of components in \linkS4class{fmx} object
}
\details{
Using definitions as S3 method dispatch \code{`[.fmx`} won't work
for \linkS4class{fmx} objects.
}
\examples{

(d = fmx('norm', mean = c(1, 4, 7), w = c(1, 1, 1)))
d[1:2]

}
\keyword{internal}
