% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{vcov.fmx}
\alias{vcov.fmx}
\title{Variance-Covariance of \linkS4class{fmx} Object}
\usage{
\method{vcov}{fmx}(object, internal = FALSE, ...)
}
\arguments{
\item{object}{\linkS4class{fmx} object}

\item{internal}{\link[base]{logical} scalar, either for the user-friendly parameters (`FALSE`, default)
(e.g., `mean,sd` for normal mixture, and `A,B,g,h` for Tukey \eqn{g}-and-\eqn{h} mixture), or
for the internal/unconstrained parameters (`TRUE`).}

\item{...}{place holder for S3 naming convention}
}
\value{
The function [vcov.fmx()] returns a \link[base]{matrix}.
}
\description{
..
}
\details{
The function [vcov.fmx()] returns 
the approximate asymptotic variance-covariance \link[base]{matrix} of the user-friendly parameters via delta-method (`parm = 'user'`), 
or the asymptotic variance-covariance matrix of the internal/unconstrained parameters (`parm = 'internal'`). 
When the distribution has constraints on one or more parameters, 
function [vcov.fmx()] does not return the variance/covariance involving the constrained parameters.
}
\keyword{internal}
