% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_fmx.R
\name{coef.fmx}
\alias{coef.fmx}
\title{Parameter Estimates of \linkS4class{fmx} object}
\usage{
\method{coef}{fmx}(object, internal = FALSE, ...)
}
\arguments{
\item{object}{\linkS4class{fmx} object}

\item{internal}{\link[base]{logical} scalar, either for the user-friendly parameters (\code{FALSE}, default)
(e.g., \verb{mean,sd} for normal mixture, and \verb{A,B,g,h} for Tukey \eqn{g}-and-\eqn{h} mixture), or
for the internal/unconstrained parameters (\code{TRUE}).}

\item{...}{place holder for S3 naming convention}
}
\value{
Function \code{\link[=coef.fmx]{coef.fmx()}} returns a \link[base]{numeric} \link[base]{vector}.
}
\description{
..
}
\details{
Function \code{\link[=coef.fmx]{coef.fmx()}} returns the estimates of the user-friendly parameters (\code{parm = 'user'}),
or the internal/unconstrained parameters (\code{parm = 'internal'}).
When the distribution has constraints on one or more parameters,
function \code{\link[=coef.fmx]{coef.fmx()}} does not return the estimates (which is constant \code{0}) of the constrained parameters.
}
\keyword{internal}
