% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_ext.R
\name{distArgs}
\alias{distArgs}
\title{Name(s) of Formal Argument(s) of Distribution}
\usage{
distArgs(distname)
}
\arguments{
\item{distname}{\link[base]{character} scalar, name of distribution}
}
\value{
Function \code{\link[=distArgs]{distArgs()}} returns a \link[base]{character} \link[base]{vector}.
}
\description{
To obtain the name(s) of distribution parameter(s).
}
\seealso{
\link[methods]{formalArgs}
}
\keyword{internal}
