% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_ext.R
\name{distType}
\alias{distType}
\title{Distribution Type}
\usage{
distType(type = c("discrete", "nonNegContinuous", "continuous"))
}
\arguments{
\item{type}{\link[base]{character} scalar}
}
\value{
Function \code{\link[=distType]{distType()}} returns a \link[base]{character} \link[base]{vector}.
}
\description{
..
}
\keyword{internal}
