% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{fmx-package}
\alias{fmx-package}
\title{Finite Mixture Parametrization}
\value{
Returns of each function will be documented separately.
}
\description{
..
}
\author{
Tingting Zhan
}
\keyword{internal}
