% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fmx.R
\name{fmx}
\alias{fmx}
\title{Create \linkS4class{fmx} Object for Finite Mixture Distribution}
\usage{
fmx(distname, w = 1, ...)
}
\arguments{
\item{distname}{\link[base]{character} scalar}

\item{w}{(optional) \link[base]{numeric} \link[base]{vector}.
Does not need to sum up to 1; \code{w/sum(w)} will be used internally.}

\item{...}{mixture distribution parameters.
See function \link[TukeyGH77]{dGH} for the names and default values of Tukey \eqn{g}-&-\eqn{h} distribution parameters,
or \link[stats]{dnorm} for the names and default values of normal distribution parameters.}
}
\value{
Function \code{\link[=fmx]{fmx()}} returns an \linkS4class{fmx} object.
}
\description{
To create \linkS4class{fmx} object for finite mixture distribution.
}
