% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_fmx.R
\name{logLik.fmx}
\alias{logLik.fmx}
\title{Log-Likelihood of \linkS4class{fmx} Object}
\usage{
\method{logLik}{fmx}(object, data = object@data, ...)
}
\arguments{
\item{object}{\linkS4class{fmx} object}

\item{data}{\link[base]{double} \link[base]{vector}, actual observations}

\item{...}{place holder for S3 naming convention}
}
\value{
Function \link{logLik.fmx} returns a \link[stats]{logLik} object with
an additional attribute \code{attr(,'logl')}.
}
\description{
..
}
\details{
Function \link{logLik.fmx} returns a \link[stats]{logLik} object indicating the log-likelihood.
An additional attribute \code{attr(,'logl')} indicates the point-wise log-likelihood,
to be use in Vuong's closeness test.
}
