% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{logLik.fmx}
\alias{logLik.fmx}
\title{Log-Likelihood of \linkS4class{fmx} Object}
\usage{
\method{logLik}{fmx}(object, ...)
}
\arguments{
\item{object}{\linkS4class{fmx} object}

\item{...}{additional parameters, currently of no use}
}
\value{
Function \code{\link[=logLik.fmx]{logLik.fmx()}} returns a \link[stats]{logLik} object.
}
\description{
Log-likelihood of an \linkS4class{fmx} object.
}
\keyword{internal}
