% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment.R
\name{moment_fmx}
\alias{moment_fmx}
\title{Moment of Each Component in an \linkS4class{fmx} Object}
\usage{
moment_fmx(object)
}
\arguments{
\item{object}{an \linkS4class{fmx} object}
}
\value{
Function \code{\link[=moment_fmx]{moment_fmx()}} returns a \link[param2moment:moment-class]{moment} object.
}
\description{
To find moments of each component in an \linkS4class{fmx} object.
}
\details{
Function \code{\link[=moment_fmx]{moment_fmx()}} calculates the \link[param2moment:moment-class]{moment}s
and distribution characteristics of each mixture component of
an S4 \linkS4class{fmx} object.
}
\examples{
(d2 = fmx('GH', A = c(1,6), B = 2, g = c(0,.3), h = c(.2,0), w = c(1,2)))
moment_fmx(d2)

}
