% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{nobs.fmx}
\alias{nobs.fmx}
\title{Number of Observations in \linkS4class{fmx} Object}
\usage{
\method{nobs}{fmx}(object, ...)
}
\arguments{
\item{object}{\linkS4class{fmx} object}

\item{...}{additional parameters, currently of no use}
}
\value{
Function \code{\link[=nobs.fmx]{nobs.fmx()}} returns an \link[base]{integer} scalar.
}
\description{
Number of observations in an \linkS4class{fmx} object.
}
\details{
Function \code{\link[=nobs.fmx]{nobs.fmx()}} finds the sample size of \verb{@data} slot of
an \linkS4class{fmx} object.
}
\keyword{internal}
