% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{nobs.fmx}
\alias{nobs.fmx}
\title{Number of Observations in \linkS4class{fmx} Object}
\usage{
\method{nobs}{fmx}(object, ...)
}
\arguments{
\item{object}{\linkS4class{fmx} object}

\item{...}{place holder for S3 naming convention}
}
\value{
Function \code{\link[=nobs.fmx]{nobs.fmx()}} returns an \link[base]{integer} scalar.
}
\description{
..
}
\details{
Function \code{\link[=nobs.fmx]{nobs.fmx()}} returns the sample size of
an \linkS4class{fmx} object
}
