% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_fmx.R
\name{print.fmx}
\alias{print.fmx}
\title{S3 \link[base]{print} of \linkS4class{fmx} Object}
\usage{
\method{print}{fmx}(x, ...)
}
\arguments{
\item{x}{an \linkS4class{fmx} object}

\item{...}{additional parameters, not currently in use}
}
\value{
Function \code{\link[=print.fmx]{print.fmx()}} returns the input \linkS4class{fmx} object invisibly.
}
\description{
..
}
\keyword{internal}
