% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmx.R
\name{qfmx_interval}
\alias{qfmx_interval}
\title{Obtain \code{interval} for \link[TukeyGH77]{vuniroot2} for Function \code{\link[=qfmx]{qfmx()}}}
\usage{
qfmx_interval(
  dist,
  p = c(1e-06, 1 - 1e-06),
  distname = dist@distname,
  K = dim(pars)[1L],
  pars = dist@pars,
  w = dist@w,
  ...
)
}
\arguments{
\item{dist}{\linkS4class{fmx} object}

\item{p}{\link[base]{length}-2 \link[base]{numeric} \link[base]{vector}}

\item{distname, K, pars, w}{(optional) ignored if \code{dist} is provided}

\item{...}{additional parameters, currently not used}
}
\value{
Function \code{\link[=qfmx_interval]{qfmx_interval()}} returns a \link[base]{length}-2 \link[base]{numeric} \link[base]{vector}.
}
\description{
Obtain \code{interval} for \link[TukeyGH77]{vuniroot2} for Function \code{\link[=qfmx]{qfmx()}}
}
\keyword{internal}
