% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_fmx.R
\name{show,fmx-method}
\alias{show,fmx-method}
\title{Show \linkS4class{fmx} Object}
\usage{
\S4method{show}{fmx}(object)
}
\arguments{
\item{object}{an \linkS4class{fmx} object}
}
\value{
The \link[methods]{show} method for \linkS4class{fmx} object
does not have a returned value.
}
\description{
Print the parameters of an \linkS4class{fmx} object and plot its density curves.
}
\keyword{internal}
