% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0S4.R
\name{[.fmx}
\alias{[.fmx}
\title{Subset of Components in \linkS4class{fmx} Object}
\usage{
\method{[}{fmx}(x, i)
}
\arguments{
\item{x}{\linkS4class{fmx} object}

\item{i}{\link[base]{integer} or \link[base]{logical} \link[base]{vector},
the \emph{row} indices of \emph{components} to be chosen, see \link[base]{[}}
}
\value{
An \linkS4class{fmx} object consisting of a subset of components.
Information about the observations (e.g. slots \verb{@data} and \verb{@data.name}),
will be lost.
}
\description{
Taking subset of components in \linkS4class{fmx} object
}
\keyword{internal}
